refactor cli gateway
