+++ hpecp/cli/base.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 # (C) Copyright [2020] Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
